Add generate_cfg.py
This commit adds a new files, `generate_cfg.py` that can
be used to generate a set of rules for a CFG based on any number
of input sentences. At the moment, it requires the Stanford parser
to be set up to work with Python, as well as some relevant parser
files.