Update CFG generator to output full ruleset
This commit builds out the CFG generator to output a text
file of all the rules for our grammar based on our corpus.
The output file is very large, and it takes a couple hours
(roughly) to generate, so this should be interesting. This
commit also adds a couple more file exts to the .gitignore.